Reorganized Draggable Repl component
