+++ src/commons/mobileWorkspace/DraggableRepl.tsx
@@ -1,11 +1,13 @@
 import React from 'react';
 import Draggable, { DraggableEventHandler } from 'react-draggable';
 
+import Repl, { ReplProps } from '../repl/Repl';
+
 type DragReplProps = {
   position: { x: number; y: number };
   onDrag: DraggableEventHandler;
-  body: JSX.Element;
   disabled: boolean;
+  replProps: ReplProps;
 };
 
 const DraggableRepl: React.FC<DragReplProps> = props => {
@@ -32,7 +34,9 @@ const DraggableRepl: React.FC<DragReplProps> = props => {
             ))}
           </div>
         )}
-        <div className="REPL-content">{props.body}</div>
+        <div className="REPL-content">
+          <Repl {...props.replProps} />
+        </div>
       </div>
     </Draggable>
   );
